+++ src/nanobot_runtime/services/hooks/ltm/consolidator.py
